Refactor: Split phone auth context into smaller files
This commit refactors the phone auth context into smaller, more manageable files to improve code organization and maintainability.
