+++ src/services/auth/sessionAuth.ts
@@ -1,10 +1,9 @@
-
 import { supabase } from '@/integrations/supabase/client';
-import { sonnerToast } from 'sonner';
+import { toast } from 'sonner';
 
 export const signOut = async () => {
   await supabase.auth.signOut();
-  sonnerToast.info('You have been signed out.');
+  toast.info('You have been signed out.');
 };
 
 export const getCurrentSession = async () => {
@@ -16,4 +15,3 @@ export const getCurrentUser = async () => {
   const { data } = await supabase.auth.getUser();
   return data.user;
 };
-
